Make the code work for mbed-coap-example
The main thing that makes the code work is the "return 0" at the
beginning of 'platform_timer_get_remaining_slots'. Obviously this
is not ideal and needs to be investigated further.
It was also observer that adding .tolerance(0) when scheduling
the timer callback makes the code much more stable (thus working
much better) than relying on the default (non 0) tolerance provided
by MINAR.